+++ packages/app/rollup.config.js
@@ -47,6 +47,10 @@ export default {
 		}),
 		replace({
 			'process.env.NODE_ENV': JSON.stringify('development'),
+			'process.env.NETWORK': JSON.stringify('devnet'),
+			'process.env.PROGRAM_ID': JSON.stringify(
+				'D47wvD2bTDXR9XqqHdP8bwYSXu2QPMW6fGHg2aEBKunM'
+			),
 		}),
 	],
 };