+++ packages/sure-sdk/src/core/constants.ts
@@ -1,7 +1,0 @@
-import { PublicKey } from '@solana/web3.js';
-
-export const PROGRAM_ID_STR = '6JqrqLaHQu9YPpp5z9KzF2nUeR4tnm3y3z9UMzkhVAwG';
-
-export const PROGRAM_ID = new PublicKey(
-	'6JqrqLaHQu9YPpp5z9KzF2nUeR4tnm3y3z9UMzkhVAwG'
-);