+++ packages/sure-sdk/src/utils/index.ts
@@ -1,3 +1,4 @@
 export { Bitmap } from './bitmap';
 export { Money } from './money';
-export { SureDate } from './time';
+export { SureDate, getUnixTime } from './time';
+export * from './transaction';