+++ packages/sure-sdk/src/utils/money.ts
@@ -1,4 +1,6 @@
 import * as anchor from '@project-serum/anchor';
+import { getMint } from '@solana/spl-token';
+import { PublicKey } from '@solana/web3.js';
 export class Money {
 	protected amount: number;
 	protected decimals: number;
@@ -12,7 +14,7 @@ export class Money {
 		return new Money(decimals, amount);
 	}
 
-	setAmount(amount: number): Money {
+	public setAmount(amount: number): Money {
 		this.amount = amount;
 		return this;
 	}
@@ -26,5 +28,29 @@ export class Money {
 	convertToAmount(amountDecimals: anchor.BN): void {
 		const decimalsBN = new anchor.BN(10).pow(new anchor.BN(this.decimals));
 		this.amount = amountDecimals.div(decimalsBN).toNumber();
+	}
+
+	static async convertBNFromDecimals(
+		connection: anchor.web3.Connection,
+		amount: anchor.BN,
+		mint: PublicKey
+	): Promise<string> {
+		const mintInfo = await getMint(connection, mint);
+		const base = new anchor.BN(10 ** mintInfo.decimals);
+		let fraction = amount.mod(base).toString(10);
+		while (fraction.length < mintInfo.decimals) {
+			fraction = `0${fraction}`;
+		}
+		const whole = amount.div(base).toString(10);
+		return `${whole}${fraction == '0' ? '' : `.${fraction}`}`;
+	}
+
+	static async convertBNToDecimals(
+		connection: anchor.web3.Connection,
+		amount: anchor.BN,
+		mint: PublicKey
+	): Promise<anchor.BN> {
+		const mintInfo = await getMint(connection, mint);
+		return amount.mul(new anchor.BN(10 ** mintInfo.decimals));
 	}
 }