+++ packages/sure-sdk/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.9 (2022-06-17)
+
+**Note:** Version bump only for package @surec/sdk
+
+
+
+
+
+## 0.0.8 (2022-06-17)
+
+**Note:** Version bump only for package @surec/sdk
+
+
+
+
+
 ## 0.0.7 (2022-06-10)
 
 **Note:** Version bump only for package @surec/sdk